requirements.txt
PyQt6>=6.6
numpy>=1.26
opencv-python>=4.9
Pillow>=10.2
scikit-learn>=1.4
scikit-image>=0.22
noise>=1.2
pyswarms>=1.3
# Optional – uncomment for neural fitness
# torch>=2.2
# torchvision>=0.17
